Changed the folder Structure and some files
