+++ Project UBORIC.md
@@ -3,3 +3,5 @@ Group members
 2.Rohit Singh              --- fw18_0393
 3.Om Kundu                 --- fw17_0736
 4.Surag VC                 --- fp03_199
+
+We are four members of team working towords cloning an e-Commerce website called UBORIC.com.
