+++ air_tight.html
@@ -7,6 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Air Tight Containers</title>
     <script src="https://kit.fontawesome.com/2a76efaedf.js" crossorigin="anonymous"></script>
+    <link rel="shortcut icon" href="https://i.im.ge/2022/06/14/rvnfcm.png" type="image/x-icon">
     <link rel="stylesheet" href="styles/airtight.css">
     <link rel="stylesheet" href="styles/footer.css">
 </head>
