+++ cart.html
@@ -1,55 +1,59 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="cart.css">
+    <title>Cart Page</title>
+    <link rel="shortcut icon" href="https://i.im.ge/2022/06/14/rvnfcm.png" type="image/x-icon">
+    <link rel="stylesheet" href="styles/cart.css">
 </head>
+
 <body>
     <div id="header"></div>
     <div id="product_div"></div>
 
 
     <!-- ------------------------------------ -->
-   
-        <h1>Cart</h1>
-        <p id="home">Home</p>
-        <p id="greater_logo">></p>
-        <p id="cart_logo">Cart</p>
 
-        <div id="amazing_offer">
-            <p>Amazing Offers On Checkout</p>
-        </div>
+    <h1>Cart</h1>
+    <p id="home">Home</p>
+    <p id="greater_logo">></p>
+    <p id="cart_logo">Cart</p>
 
-        <div id="left_product_div">
+    <div id="amazing_offer">
+        <p>Amazing Offers On Checkout</p>
+    </div>
+
+    <div id="left_product_div">
         <div id="left2">
             <p>Product</p>
             <p>Price</p>
             <p>Quantity</p>
             <p>Subtotal</p>
-        </div>   
-           
-           <div id="line"></div>
-           <div id="line2"></div>
-           <div id="line3"></div>
-           <div id="line4">
-           
-           </div>
-           
-        </div>
-        <div id="right_product_div">
-            <h1>Cart totals</h1>
-
-            <h3 id="subtotal">Subtotal</h3>
-            <div id="price_div"></div>
-            <div id="line1"></div>
-            <button id="proceed_button">Proceed To Checkout</button>
         </div>
 
+        <div id="line"></div>
+        <div id="line2"></div>
+        <div id="line3"></div>
+        <div id="line4">
 
-    
+        </div>
+
+    </div>
+    <div id="right_product_div">
+        <h1>Cart totals</h1>
+
+        <h3 id="subtotal">Subtotal</h3>
+        <div id="price_div"></div>
+        <div id="line1"></div>
+        <button id="proceed_button">Proceed To Checkout</button>
+    </div>
+
+
+
 </body>
+
 </html>
-<script src="./cart.js" type="module"></script>
+<script src="scripts/cart.js" type="module"></script>
