+++ index.html
@@ -1,17 +1,97 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Best Online Shopping Sites in India</title>
-    <link rel="stylesheet" href="index.css">
+    <title>Best Online Shopping Sites in India | Shop for clothing, shoes, grocery, electric, Home appliances and more.</title>
+    <link rel="shortcut icon" href="https://i.im.ge/2022/06/14/rvnfcm.png" type="image/x-icon">
+    <script src="https://kit.fontawesome.com/2a76efaedf.js" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="styles/homepage.css">
+    <link rel="stylesheet" href="styles/footer.css">
+    <link rel="stylesheet" href="styles/navbar.css">
 </head>
+
 <body>
-    <div id="header"></div>
-    <div id="product_div"></div>
-    
-  
+    <div id="navbar"></div>
+    <div id="product_navbar"></div>
+    <div id="animated_div"></div>
+    <div id="main">
+        <div id="kitchenware">
+            <div><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/kitchen-ware-heading-pkx8tvxfmbbsdlrzrk3iwlbxruc7cn5sq339xkx9cg.png" alt=""></div>
+            <div>
+                <h2>KITCHEN WARE</h2>
+            </div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/kitchen-ware-heading-pkx8tvxfmbbsdlrzrk3iwlbxruc7cn5sq339xkx9cg.png" alt=""></div>
+        </div>
+        <div id="underkitchen">
+            <div id="und1">
+                <a href="kitchentools.html"><img src="https://www.uboric.com/wp-content/uploads/2020/12/kithchen-tools-768x602.jpg" alt=""></a>
+            </div>
+            <div id="und2">
+                <a href="air_tight.html"><img src="https://www.uboric.com/wp-content/uploads/2020/12/Air-Tight-Containers-768x602.jpg" alt=""></a>
+            </div>
+            <div id="und3">
+                <a href="gaslighter.html"><img src="https://www.uboric.com/wp-content/uploads/2020/12/Gas-Lighter-1024x1024.jpg" alt=""></a>
+            </div>
+        </div>
+        <div id="info">
+            <img src="https://www.uboric.com/wp-content/uploads/2020/08/Uboric-1.jpg" alt="">
+        </div>
+        <div id="othercat">
+            <div id="oth1"><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/other-category-heading-pkx8tu1r8n97qduq2ja9rlt0l2lgx8yc1tsaz101ow.png" alt=""></div>
+            <div id="oth2">OTHER CATEGORIES</div>
+            <div id="oth3"><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/other-category-heading-2-e1608267154124-pkx8tt3x1t7xerw380vn741jzoq3pjulpp4thr1fv4.png" alt=""></div>
+        </div>
+        <div id="two_img">
+            <div><img src="https://www.uboric.com/wp-content/uploads/2021/12/banner1-1-1.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2021/12/banner2-1.jpg" alt=""></div>
+        </div>
+        <div id="head_phone">
+            <img src="https://www.uboric.com/wp-content/uploads/2021/02/Headphone.jpg" alt="">
+        </div>
+        <div id="two_img2">
+            <div><img src="https://www.uboric.com/wp-content/uploads/2021/12/banner3-1.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2021/12/banner4-1.jpg" alt=""></div>
+        </div>
+        <div id="add_grid">
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Socks-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Personal-Care-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Braclate-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Grocery-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Health-Care-400x601.jpg" alt=""></div>
+        </div>
+        <div id="spices">
+            <div id="spi1"><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/spices-heading-2-e1608267173425-pkx8te2i0gnc90hxnudm37u6his8ae6wbmp1tbnqmo.png" alt=""></div>
+            <div id="spi2">SPICES</div>
+            <div id="spi3"><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/spices-heading-pkx8tf0c7aomkmgkics8npln2wnli3amnrcjalmcgg.png" alt=""></div>
+        </div>
+        <div id="spices_grid">
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Spices-1-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Spices-2-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Spices-3-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Spices-5-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Spices-4-400x601.jpg" alt=""></div>
+        </div>
+        <div id="footwear">
+            <div id="foot1"><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/shoes-icon-right-e1605175492472-pkxdahmpc96i32hqyh9qfkc7owqxivqr4zuwukso7k.png" alt=""></div>
+            <div id="foot2">FOOT WEAR</div>
+            <div id="foot3"><img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/shoes-icon-left-e1605175515419-pkxdaq391ri2zk5gl2xdk07d1dl8g5oc65qa62g4nk.png" alt=""></div>
+        </div>
+        <div id="foot_wear_add">
+            <img src="https://www.uboric.com/wp-content/uploads/elementor/thumbs/men-fashion-pf2rgkwglk7z47uh031vpxcshl64adc389ivxapmss.jpg" alt="">
+        </div>
+        <div id="footwear_gridbox">
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Foot-Wear-1-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Foot-Wear-3-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Foot-Wear-2-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Foot-Wear-400x601.jpg" alt=""></div>
+            <div><img src="https://www.uboric.com/wp-content/uploads/2020/12/Foot-Wear-5-400x601.jpg" alt=""></div>
+        </div>
+    </div>
+    <div id="footer"></div>
 </body>
+
 </html>
-<script src="index.js" type="module"></script>
+<script src="scripts/homepage.js" type="module"></script>
