+++ scripts/cart.js
@@ -1,0 +1,7 @@
+import { navbar } from '../components/navbar.js'
+let navbar_div = document.getElementById("header")
+navbar_div.innerHTML = navbar();
+
+import { product_nav } from '../components/navbar.js'
+let lower_div = document.getElementById("product_div")
+lower_div.innerHTML = product_nav()
