+++ signup.html
@@ -1,32 +1,37 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="signup.css">
+    <title>Sign up</title>
+    <link rel="shortcut icon" href="https://i.im.ge/2022/06/14/rvnfcm.png" type="image/x-icon">
+    <link rel="stylesheet" href="styles/signup.css">
 </head>
+
 <body>
     <div id="header"></div>
     <div id="product_div"></div>
 
-
     <div id="signup_div">
         <h1>Sign Up</h1>
-       <a href="user.html"><p id="sign_in">CLEAR HEAR TO LOGIN</p></a> 
-       <form>
-        <input id="name" placeholder="First-Name" type="text" required/>
-     <br/>
-     <input id="last_name" placeholder="Last-Name" type="text" required/>
-     <br/>
-     <input id="email" placeholder="Email-Address" type="text" required/>
-     <br/>
-     <input id="password" placeholder="Password" type="password" required/>
-    </br>
-        <button id="login">REGISTER</button>
-       </form>
+        <a href="user.html">
+            <p id="sign_in">CLEAR HEAR TO LOGIN</p>
+        </a>
+        <form>
+            <input id="name" placeholder="First-Name" type="text" required/>
+            <br/>
+            <input id="last_name" placeholder="Last-Name" type="text" required/>
+            <br/>
+            <input id="email" placeholder="Email-Address" type="text" required/>
+            <br/>
+            <input id="password" placeholder="Password" type="password" required/>
+            </br>
+            <button id="login">REGISTER</button>
+        </form>
     </div>
 </body>
+
 </html>
-<script src="./signup.js" type="module"></script>
+<script src="scripts/signup.js" type="module"></script>
