+++ styles/payment.css
@@ -1,0 +1,78 @@
+.row {
+     
+    display: flex;
+    
+    
+    margin: 0 -20px;
+  }
+  
+  .col-25 {
+    
+    flex: 20%;
+  }
+  
+  .col-50 {
+    
+    flex: 10%;
+  }
+  
+  .col-75 {
+     
+    flex: 65%;
+  }
+  
+  .col-25,
+  .col-50,
+  .col-75 {
+    padding: 0 20px;
+  }
+  
+  .container {
+    background-color: #f1f7ea;
+    padding: 5px 20px 15px 20px;
+    border: 1px solid rgb(75, 42, 42);
+    border-radius: 3px;
+  }
+  
+  input[type=text] {
+    width: 70%;
+    margin-bottom: 20px;
+    padding: 10px;
+    border: 1px solid rgb(216, 207, 207);
+    border-radius: 3px;
+  }
+  
+  label {
+    margin-bottom: 10px;
+    display: block;
+  }
+  
+  .icon-container {
+    margin-bottom: 20px;
+    padding: 7px 0;
+    font-size: 24px;
+  }
+  
+  .btn {
+    background-color: #0a0c0b;
+    color: white;
+    padding: 12px;
+    margin: 10px 0;
+    border: none;
+    width: 100%;
+    border-radius: 3px;
+    cursor: pointer;
+    font-size: 17px;
+  }
+  
+  .btn:hover {
+    background-color: #e92b12;
+  }
+  
+  span.price {
+    float: right;
+    color: grey;
+  }
+  
+ 
+
