+++ user.html
@@ -1,32 +1,38 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="user.css">
+    <title>Log in</title>
+    <link rel="shortcut icon" href="https://i.im.ge/2022/06/14/rvnfcm.png" type="image/x-icon">
+    <link rel="stylesheet" href="styles/user.css">
 </head>
+
 <body>
     <div id="header">
-    </div> 
+    </div>
     <div id="product_div"></div>
     <div id="signup_div">
         <h1>Sign in</h1>
-       <a href="signup.html"><p id="sign_in">CREATE AN ACCOUNT</p></a> 
-       <form>
-        <p id="username">Username or email <img id="star" src="./images/star.png" alt=""></p>
-        <input id="email"  type="email" required>
-        <p id="Password">Password <img id="star" src="./images/star.png" alt=""></p>
-        <input id="password"  type="password" required>
+        <a href="signup.html">
+            <p id="sign_in">CREATE AN ACCOUNT</p>
+        </a>
+        <form>
+            <p id="username">Username or email <img id="star" src="./images/star.png" alt=""></p>
+            <input id="email" type="email" required>
+            <p id="Password">Password <img id="star" src="./images/star.png" alt=""></p>
+            <input id="password" type="password" required>
 
-        <p id="lost_your_password" >LOST  YOUR PASSWORD?</p>
-        <button id="login">LOGIN</button>
-       </form>
+            <p id="lost_your_password">LOST YOUR PASSWORD?</p>
+            <button id="login">LOGIN</button>
+        </form>
     </div>
 
 
-    
+
 </body>
+
 </html>
-<script src="/user.js" type="module"></script>
+<script src="scripts/user.js" type="module"></script>
